+++ src/components/pages/ComicsPage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ComicsList from "../comicsList/ComicsList";
+import AppBanner from "../appBanner/AppBanner";
+
+const ComicsPage = () => {
+    return (
+        <>
+            <AppBanner />
+            <ComicsList />
+        </>
+    );
+};
+
+export default ComicsPage;
